+++ src/server/lib/index.js
@@ -1,0 +1,10 @@
+'use strict';
+
+import { Sites, Sections } from './constant';
+import { StripHost, RemoveExtraSpace } from './parse';
+import Get from './get';
+import Catch from './route';
+
+module.exports = {
+  Sites, Sections, StripHost, RemoveExtraSpace, Get, Catch
+};